requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=10.0.0
django-filter>=23.3
django-widget-tweaks>=1.5.0